fix inaccurate sleep between pulses (#3)
This massively improves the precision of the sleeps that separate
pulses. See the doc block on `sleepFor`. With this patch the sleep after
each transmission becomes obsolete as the outlets not reacting to code
transmissions was caused by the inaccurate sleep durations between
pulses and not by tight packing of the sequences sent.